feat(resource-service): Compute git auth method once per API request
