+++ resource-service/handler/project_manager_test.go
@@ -15,6 +15,16 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
+var credentials = common_models.GitCredentials{
+	User: "my-user",
+	HttpsAuth: &apimodels.HttpsGitAuth{
+		Token: "my-token",
+	},
+	RemoteURL: "my-remote-uri",
+}
+
+var auth, _ = getAuthMethod(&credentials)
+
 type projectManagerTestFields struct {
 	git              *common_mock.IGitMock
 	credentialReader *common_mock.CredentialReaderMock
@@ -27,14 +37,9 @@ func TestProjectManager_CreateProject(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -74,14 +79,9 @@ func TestProjectManager_CreateProject_ProjectAlreadyExists(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -146,14 +146,9 @@ func TestProjectManager_CreateProject_ProjectRepoDoesNotExist(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -190,14 +185,9 @@ func TestProjectManager_CreateProject_WritingFileFails(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -234,14 +224,9 @@ func TestProjectManager_CreateProject_CommitFails(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -278,14 +263,9 @@ func TestProjectManager_UpdateProject(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -326,14 +306,9 @@ func TestProjectManager_UpdateProject_WithMigration(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -383,14 +358,9 @@ func TestProjectManager_UpdateProject_WithMigration_CannotPull(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -444,14 +414,9 @@ func TestProjectManager_UpdateProject_WithMigration_MigrationFailsOnFirstTry(t *
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -510,14 +475,9 @@ func TestProjectManager_UpdateProject_WithMigration_AlreadyMigrated(t *testing.T
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -568,14 +528,9 @@ func TestProjectManager_UpdateProject_WithMigration_InvalidMetadata(t *testing.T
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -627,14 +582,9 @@ func TestProjectManager_UpdateProject_WithMigration_NoMetadata(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -714,14 +664,9 @@ func TestProjectManager_UpdateProject_ProjectDoesNotExist(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -752,14 +697,9 @@ func TestProjectManager_UpdateProject_ProjectNotInitialized(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -793,14 +733,9 @@ func TestProjectManager_UpdateProject_ProjectNotInitializedEmptyMetadataFile(t *
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -834,14 +769,9 @@ func TestProjectManager_UpdateProject_CannotGetDefaultBranch(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
@@ -878,14 +808,9 @@ func TestProjectManager_UpdateProject_CheckoutBranchFails(t *testing.T) {
 	}
 
 	expectedGitContext := common_models.GitContext{
-		Project: "my-project",
-		Credentials: &common_models.GitCredentials{
-			User: "my-user",
-			HttpsAuth: &apimodels.HttpsGitAuth{
-				Token: "my-token",
-			},
-			RemoteURL: "my-remote-uri",
-		},
+		Project:     "my-project",
+		Credentials: &credentials,
+		AuthMethod:  auth,
 	}
 
 	fields := getTestProjectManagerFields()
